+++ app/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_executable(matapos src/main.cpp)
+
+target_link_libraries(matapos barcodereader window camera)
