+++ app/src/main.cpp
@@ -1,14 +1,10 @@
-#include "camera/camera.h"
-#include "barcodereader/barcodereader.h"
-
-#include "guiheaderfiles/window.h"
-
 #include <QApplication>
+#include "window.h"
+
 
 int main(int argc, char* argv[]) {
-	
+
 	QApplication app(argc, argv);
-	
 	Window window;
 	
 	window.show();
