--- include/simplecameralib/printhello.h
@@ -1,6 +1,0 @@
-#ifndef HELLO_PRINTER
-#define HELLO_PRINTER
-
-void printhello();
-
-#endif
