+++ lib/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_subdirectory(barcodereader)
+add_subdirectory(camera)
+add_subdirectory(window)
+
