+++ lib/barcodereader/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(barcodereader src/barcodereader.cpp include/barcodereader.h)
+
+target_include_directories(barcodereader PUBLIC include/)
+
+target_link_libraries(barcodereader camera Threads::Threads PkgConfig::ZBar ${OpenCV_LIBS})
