+++ lib/barcodereader/include/barcodereader.h
