+++ lib/camera/include/camera.h
