+++ lib/database/src/database.cpp
@@ -1,4 +1,4 @@
-#include "db/database.h"
+#include "database.h"
 
 Database::Database() {
