--- src/CMakeLists.txt
@@ -1,4 +1,0 @@
-cmake_minimum_required (VERSION 3.8)
-
-project(matapos)
-
