--- src/printhello.cpp
@@ -1,7 +1,0 @@
-#include <iostream>
-#include "simplecameralib/printhello.h"
-
-
-void printhello() {
-	std::cout << "Hello folks \n";
-}
